chore: minor cleanups + remove util fn
